+++ Control/Joint/Abilities/Adaptable.hs
@@ -1,7 +1,8 @@
 module Control.Joint.Abilities.Adaptable where
 
 import Control.Joint.Core (type (~>))
+import Control.Joint.Abilities.Liftable (Liftable (lift))
+import Control.Joint.Abilities.Transformer ((:>) (T))
 
 class Adaptable (subeff :: * -> *) (eff :: * -> *) | subeff -> eff where
-	{-# MINIMAL adapt #-}
 	adapt :: subeff ~> eff